feat(portfolio): Add ability to unpublish
Project showcase can be unpublished without deleting the db entry.
This enables a project to be temporarily made non visible on the
website.

Add `.publish` field to `Project` model and only list projects that
have the field set to `True`.
Detail page of `Project` object with `.publish` set to `False` return
404 status.